Enhancements in views.py and creation of base.html in templates folder
The exception in views.py was modified to be more specific with an IntegrityError. In the templates folder, a base.html was created to establish the foundation of the code that will be repeated on all other pages, where for now, the navbar is located. It was also added that after registering a new user, a cookie with the session ID is stored.